+++ ClubEconomyControl/Models/Interfaces/IEconomyModel.cs
@@ -3,7 +3,7 @@
     public interface IEconomyModel
     {
         int Id { get; set; }
-        int Amount { get; set; }
+        decimal Amount { get; set; }
         string? Description { get; set; }
         int ClubId { get; set; }
     }
